update epoch set to fif (problem documented in comments)
Problem: The .fif epoch that is saved is147MB big (compared to the 1.4-1.9GB of the .set). This massive reduction in size is suspicious. But loading and visualising the .fif reveals that it has the same data as the .set file.
Also suspicious: the .fif of all subj have the same size. But loading and plotting reveals different data