MVPA & PreStim: Lots of updates
PreStim is now running for freqbands - but. There seems to be a bug. The pValues for eeg_data are 1 for every timepoint and channel. Has to be a bug. Running now with a different solver than lbfgs to see if that's the issue, because the problem didn't appear until I changed it from default (newton) to lbfgs.