LogReg: Attempt to get the permutation test to work (failed)
Some issue on line 673 in the pymer4/models/Lmer package, which can be fixed by 'perm_dat = pandas2R(perm_dat)' - but then another error occurs in pymer4/models/Lmer line 675, because of pymer4/utils line 620: the t-values cannot be returned because they are of type tuple and not integer. I should do an issue report on that on the pymer4's github.

This commit is just so I have a record of my script before I undo the changes that are unnecessary if the permuting doesn't work.